Move versions from internal packages
