+++ src/test/scala/com/eclipsesource/schema/ErrorReportingSpec.scala
@@ -1,14 +1,14 @@
 package com.eclipsesource.schema
 
+import com.eclipsesource.schema.drafts.Version4
 import com.eclipsesource.schema.internal.Keywords
-import com.eclipsesource.schema.internal.draft4.Version4
 import com.eclipsesource.schema.internal.serialization.SchemaReads
 import org.specs2.mutable.Specification
 import play.api.libs.json._
 
 class ErrorReportingSpec extends Specification {
 
-  import com.eclipsesource.schema.internal.draft4.Version4._
+  import Version4._
 
   case class Post(id: Long, title: String, body: String)
