+++ src/test/scala/com/eclipsesource/schema/ExclusiveMaximumSpec.scala
@@ -1,6 +1,6 @@
 package com.eclipsesource.schema
 
-import com.eclipsesource.schema.internal.draft7.Version7
+import com.eclipsesource.schema.drafts.Version7
 import com.eclipsesource.schema.test.JsonSpec
 import org.specs2.mutable.Specification
