+++ src/test/scala/com/eclipsesource/schema/FormatSpec.scala
@@ -1,6 +1,6 @@
 package com.eclipsesource.schema
 
-import com.eclipsesource.schema.internal.draft4.Version4
+import com.eclipsesource.schema.drafts.Version4
 import com.eclipsesource.schema.internal.validators.DefaultFormats.DatetimeFormat
 import org.specs2.mutable.Specification
 import com.eclipsesource.schema.test.JsonSpec
