+++ src/test/scala/com/eclipsesource/schema/Issue99Spec.scala
@@ -1,14 +1,14 @@
 package com.eclipsesource.schema
 
-import com.eclipsesource.schema.internal.draft4.Version4
+import com.eclipsesource.schema.drafts.{Version4, Version7}
 import org.specs2.mutable.Specification
 import play.api.libs.json.Json
 
-class Issue99Spec extends Specification {
-
-  import com.eclipsesource.schema.internal.draft4.Version4._
+class Issue99Spec extends Specification { self =>
 
   "Issue 99 Spec" should {
+
+    import Version4._
 
     "validate issue 99 test case" in {
       val schema = JsonSource.schemaFromString(
@@ -33,7 +33,7 @@ class Issue99Spec extends Specification {
     }
 
     "validate issue 99-1 test case via URL" in {
-      val schemaUrl = getClass.getResource("/issue-99-1.json")
+      val schemaUrl = self.getClass.getResource("/issue-99-1.json")
       val validator = SchemaValidator(Some(Version4))
       val result = validator.validate(schemaUrl, Json.obj(
         "mything" -> "test"
@@ -42,7 +42,7 @@ class Issue99Spec extends Specification {
     }
 
     "not cause stack overflow for issue 99-5 test case via URL" in {
-      val schemaUrl = getClass.getResource("/issue-99-5.json")
+      val schemaUrl = self.getClass.getResource("/issue-99-5.json")
       val validator = SchemaValidator(Some(Version4))
       // must terminate
       validator.validate(schemaUrl, Json.obj(
