+++ src/test/scala/com/eclipsesource/schema/NotSpec.scala
@@ -1,10 +1,8 @@
 package com.eclipsesource.schema
 
+import com.eclipsesource.schema.drafts.{Version4, Version7}
 import com.eclipsesource.schema.test.JsonSpec
 import org.specs2.mutable.Specification
-
-import com.eclipsesource.schema.internal.draft4.Version4
-import com.eclipsesource.schema.internal.draft7.Version7
 
 class NotSpec extends Specification with JsonSpec {
