+++ src/test/scala/com/eclipsesource/schema/UniqueItemsSpec.scala
@@ -1,7 +1,6 @@
 package com.eclipsesource.schema
 
-import com.eclipsesource.schema.internal.draft4.Version4
-import com.eclipsesource.schema.internal.draft7.Version7
+import com.eclipsesource.schema.drafts.{Version4, Version7}
 import com.eclipsesource.schema.test.JsonSpec
 import org.specs2.mutable.Specification
