+++ src/test/scala/com/eclipsesource/schema/internal/refs/ResolveNumberConstraintsSpec.scala
@@ -1,6 +1,6 @@
 package com.eclipsesource.schema.internal.refs
 
-import com.eclipsesource.schema.internal.draft4.Version4
+import com.eclipsesource.schema.drafts.Version4
 import com.eclipsesource.schema.{JsonSource, SchemaType, SchemaValue}
 import org.specs2.mutable.Specification
 import play.api.libs.json.JsNumber
