+++ src/test/scala/com/eclipsesource/schema/internal/serialization/SchemaReadsSpec.scala
@@ -1,6 +1,7 @@
 package com.eclipsesource.schema.internal.serialization
 
 import com.eclipsesource.schema.{JsonSource, SchemaType, SchemaValue}
+import com.eclipsesource.schema.drafts.{Version4, Version7}
 import org.specs2.mutable.Specification
 import play.api.libs.json.{JsBoolean, Json}
 
@@ -9,7 +10,7 @@ class SchemaReadsSpec extends Specification {
 
   "Schema Reads for draft 4" should {
 
-    import com.eclipsesource.schema.internal.draft4.Version4._
+    import Version4._
 
     "not fail with match error (#104)" in {
       val schema = JsonSource.schemaFromString("""
@@ -27,7 +28,7 @@ class SchemaReadsSpec extends Specification {
 
   "Schema Reads for draft 7" should {
 
-    import com.eclipsesource.schema.internal.draft7.Version7._
+    import Version7._
 
     "read boolean schema" in {
       val booleanSchema = Json.fromJson[SchemaType](JsBoolean(true)).get
